requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
h11==0.13.0
idna==3.3
numpy==1.22.2
outcome==1.1.0
pandas==1.4.1
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
schedule==1.1.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
python-dotenv~=0.20.0
pydantic~=1.9.0